requirements.txt
yfinance
pandas
numpy
dash
dash-bootstrap-components
plotly
tensorflow
scikit-learn
transformers
torch
openai
praw
python-dotenv
arch
pmdarima
statsmodels
seaborn
matplotlib
joblib
scipy
requests
tqdm
beautifulsoup4
SQLAlchemy
pytest
pandas-datareader
pytz
keras
numpy-financial
ta
yahoofinancials
nltk
streamlit>=1.22.0